% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{CdFromDensities}
\alias{CdFromDensities}
\title{The dominance rate of X_A over X_B given the density functions.}
\usage{
CdFromDensities(densityX_A, densityX_B, xlims, EPSILON = 0.001)
}
\arguments{
\item{densityX_A}{The probability density function of the random variable X_A.}

\item{densityX_B}{The probability density function of the random variable X_B.}

\item{xlims}{an interval that represents the domain of definition the density functions.}
}
\value{
Returns the dominance rate of X_A over X_B.
}
\description{
Returns a real number in the interval [0,1] that represents the dominance rate of X_A over X_B.
Basically, we are measuring the amount of mass of X_A in which the cumulative distribution of X_A is higher minus the amount of mass of X_B in which the cumulative distribution of X_B is higher.
}
\examples{
# If two symmetric distributions are centered in the same point (x = 0 in this case), then their Cd will be 0.5.
densityX_A <- normalDensity(0,1)
densityX_B <- uniformDensity(c(-2,2))
Cd = CdFromDensities(densityX_A, densityX_B, c(-5,5))
plot(densityX_A, from=-5, to=5, type="l",  col="red", xlab="x", ylab="probability density")
curve(densityX_B, add=TRUE, col="blue", type="l", lty=2)
mtext(paste("Cd(X_A, X_B) =", format(round(Cd, 3), nsmall = 3)), side=3) # add Cd to plot as text
legend(x = c(-4.5, -2), y = c(0.325, 0.4),legend=c("X_A", "X_B"), col=c("red", "blue"), lty=1:2, cex=0.8) # add legend


### Example 2 ###
# If two distributions are equal, Cd will be 0.5.  Cd(X_A,X_A) = 0.5
CdFromDensities(densityX_A, densityX_A, c(-10,10))


### Example 3 ###
densityX_A <- normalDensity(-2,1)
densityX_B <- uniformDensity(c(-2,2))
# Cd(X_A,X_B) = 1 - Cd(X_B, X_A)
CpFromDensities(densityX_A, densityX_B, c(-Inf,Inf))
1 - CpFromDensities(densityX_B, densityX_A, c(-Inf,Inf))
}
\seealso{
\code{\link{CpFromDensities}}
}
