% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{coerceToDifferenceArea}
\alias{coerceToDifferenceArea}
\title{Coerce values to the square in the diffference graph}
\usage{
coerceToDifferenceArea(p, v)
}
\arguments{
\item{p}{the x values.}

\item{v}{the values ot be corrected.}
}
\value{
vector with the v values corrected
}
\description{
Coerce values to the square in the diffference graph
}
\keyword{internal}
