% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_helper_from_ranks_to_integrable_values}
\alias{cpp_helper_from_ranks_to_integrable_values}
\title{Helper function for from_ranks_to_integrable_values}
\usage{
cpp_helper_from_ranks_to_integrable_values(rank_interval_mult, j_max)
}
\arguments{
\item{rank_interval_mult}{the value of the normalized density of Y_A.}

\item{j_max}{maximum number of sample points.}
}
\description{
Helper function for from_ranks_to_integrable_values
}
