% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{cumulative_difference_plot}
\alias{cumulative_difference_plot}
\title{Generate the cumulative difference-plot}
\usage{
cumulative_difference_plot(
  X_A_observed,
  X_B_observed,
  isMinimizationProblem,
  labelA = "X_A",
  labelB = "X_B",
  alpha = 0.05,
  EPSILON = 1e-20,
  nOfBootstrapSamples = 1000,
  ignoreMinimumLengthCheck = FALSE
)
}
\arguments{
\item{X_A_observed}{array of the observed samples (real values) of X_A.}

\item{X_B_observed}{array of the observed samples (real values) of X_B, it needs to have the same length as X_A.}

\item{isMinimizationProblem}{a boolean value where TRUE represents that lower values are preferred to larger values.}

\item{labelA}{(optional, default value "X_A") the label corresponding to X_A.}

\item{labelB}{(optional, default value "X_B") the label corresponding to X_B.}

\item{alpha}{(optional, default value 0.05) the error of the confidence interval. If alpha = 0.05 then we have 95 percent confidence interval.}

\item{EPSILON}{(optional, default value 1e-20) minimum difference between two values to be considered different.}

\item{nOfBootstrapSamples}{(optional, default value 1e3) how many bootstrap samples to average. Increases computation time.}

\item{ignoreMinimumLengthCheck}{(optional, default value FALSE) whether to check for a minimum length in X_A and X_B.}
}
\value{
retunrs and shows the cumulative difference-plot
}
\description{
Generate the cumulative difference-plot given the observed samples using the bootstrap method.
}
\examples{

### Example 1 ###
X_A_observed <- rnorm(100, mean = 2, sd = 1)
X_B_observed <- rnorm(100, mean = 2.1, sd = 0.5)
\donttest{
 cumulative_difference_plot(X_A_observed, X_B_observed, TRUE, labelA="X_A", labelB="X_B")
}
\dontshow{
 cumulative_difference_plot(X_A_observed, X_B_observed, TRUE, labelA="X_A", labelB="X_B", nOfBootstrapSamples=100)
}
}
