% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{get_X_prima_AB_bounds_bootstrap}
\alias{get_X_prima_AB_bounds_bootstrap}
\title{Estimate X'_A and X'_B bounds with bootstrap}
\usage{
get_X_prima_AB_bounds_bootstrap(
  X_A_observed,
  X_B_observed,
  nOfQuantiles,
  nOfBootstrapSamples = 10000,
  alpha = 0.2,
  EPSILON = 1e-20,
  returnDataframe = FALSE
)
}
\arguments{
\item{X_A_observed}{array of the observed samples (real values) of X_A.}

\item{X_B_observed}{array of the observed samples (real values) of X_B.}

\item{nOfQuantiles}{the number of points in the interval [0,1] in which the density is estimated.}

\item{nOfBootstrapSamples}{(optional, default value 1e6) how many bootsrap samples to average.}

\item{alpha}{(optional, default value 0.2) the error of the confidence interval.}

\item{EPSILON}{(optional, default value 1e-20) minimum difference between two values to be considered different.}

\item{returnDataframe}{(optional, default value FALSE, as it returns a list by default) Wether to return a dataframe or a list.}
}
\value{
Returns a list with the following fields:

- p: values in the interval [0,1] that represent the nOfQuantiles points in which the densities are estimated. Useful for plotting.

- X_prima_A_cumulative_estimation: an array with the estimated probability densites of X_prima_A for each point (p[[i]] + p[[i+1]])/2.

- X_prima_A_cumulative_upper: an array with the upper bounds of confidence 1 - alpha of the density estimation of X_prima_A

- X_prima_A_cumulative_lower: an array with the lower bounds of confidence 1 - alpha of the density estimation of X_prima_A

- X_prima_B_cumulative_estimation: The same as X_prima_A_cumulative_estimation for X'_B.

- X_prima_B_cumulative_upper: The same as X_prima_A_cumulative_upper for X'_B

- X_prima_B_cumulative_lower: The same as X_prima_A_cumulative_lower for X'_B
}
\description{
Estimate the confidence intervals for the density estimations of X'_A and X'_B using bootstrap.
As a bonus, in addition to the density, the bounds of the cumulative density are also compared.
}
\examples{
library(ggplot2)

X_A_observed <- c(0,1,1,2,2,3,4)
X_B_observed <- c(0,1,1,2,3,1,1,5,6)
res <- get_X_prima_AB_bounds_bootstrap(X_A_observed, X_B_observed, 100, returnDataframe=TRUE, nOfBootstrapSamples=1e2)

densityesPlot = ggplot() +
 geom_ribbon(data = res, aes(x=p, ymin = X_prima_B_cumulative_lower, ymax = X_prima_B_cumulative_upper), fill = "#00BFC4", alpha = 0.25) +
 scale_colour_manual("", breaks = c("X'_A", "X'_B"),  values = c("#F8766D", "#00BFC4")) +
 geom_line(data = res, aes(x=p, y=X_prima_A_cumulative_estimation, colour = "X'_A", linetype="X'_A")) +
 geom_line(data = res, aes(x=p, y=X_prima_B_cumulative_estimation, colour = "X'_B",  linetype ="X'_B")) +
 geom_ribbon(data = res, aes(x=p, ymin = X_prima_A_cumulative_lower, ymax = X_prima_A_cumulative_upper), fill = "#F8766D", alpha = 0.25) +
 scale_linetype_manual("", breaks = c("X'_A", "X'_B"), values = c("dashed", "solid")) +
 xlab('x') +
 ylab('cumulative probability')
 print(densityesPlot)
}
