% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{getEmpiricalCumulativeDistributions}
\alias{getEmpiricalCumulativeDistributions}
\title{Get the empirical distribution from samples.}
\usage{
getEmpiricalCumulativeDistributions(
  X_A_observed,
  X_B_observed,
  nOfEstimationPoints,
  EPSILON
)
}
\arguments{
\item{X_A_observed}{array of the observed samples (real values) of X_A.}

\item{X_B_observed}{array of the observed samples (real values) of X_B.}

\item{nOfEstimationPoints}{the number of points in the interval [0,1] in which the cumulative density is estimated + 2.}

\item{EPSILON}{(optional, default value 1e-20) minimum difference between two values to be considered different.}
}
\value{
a list with two fields: the empirical distributions of X'A and X'B.
}
\description{
Given the observed sampels of X_A (or X_B) returns the empirical cumulative distribution
function of X'_A (or X'_B)
}
\examples{
### Example 1 ###
}
