% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{helperGet_X_prima_AB_bounds_bootstrap}
\alias{helperGet_X_prima_AB_bounds_bootstrap}
\title{Helper function for get_X_prima_AB_bounds_bootstrap.}
\usage{
helperGet_X_prima_AB_bounds_bootstrap(sortedRanks, r_max, j, j_max)
}
\arguments{
\item{sortedRanks}{the sorted ranks of either the observed X_A or X_B.}

\item{r_max}{The largest rank.}

\item{j}{the index that corresponds to the position. j goes from 0 to nDatapointsWhereDensityEstimated-1}

\item{j_max}{the largest index that will be used. its value is nDatapointsWhereDensityEstimated-1}
}
\value{
the probability density in this point
}
\description{
The density corresponding to the position in index j is computed, given the SORTED ranks, and r_max
}
\examples{
j_max <- 1000
f <- function(x){helperGet_X_prima_AB_bounds_bootstrap(sortedRanks=c(0,0,0,0,0,0,2,3,4), r_max=4, j=x, j_max=j_max)}
plot(x = 0:j_max / j_max, y = sapply(0:j_max, f), type="l")
plot(x = 0:j_max / j_max, y = cumsum(c(0,head(sapply(0:j_max, f),-1)) / j_max), type="l")
}
\keyword{internal}
