% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{helperGet_X_prima_AB_bounds_bootstrap_for_every_j}
\alias{helperGet_X_prima_AB_bounds_bootstrap_for_every_j}
\title{Helper function for get_X_prima_AB_bounds_bootstrap.}
\usage{
helperGet_X_prima_AB_bounds_bootstrap_for_every_j(sortedRanks, r_max, j_max)
}
\arguments{
\item{sortedRanks}{the sorted ranks of either the observed X_A or X_B.}

\item{r_max}{The largest rank.}

\item{j_max}{the largest index that will be used. its value is nDatapointsWhereDensityEstimated-1}
}
\value{
the probability density in this point
}
\description{
The density corresponding to the position in index j is computed, given the SORTED ranks, and r_max
}
\examples{
### Example 1 ###
j_max <- 12
r_max <- 6
sortedRanks <- c(0,0,0,0,1,1,1,1,1,1,1,3,4)
densities <- helperGet_X_prima_AB_bounds_bootstrap_for_every_j(sortedRanks=sortedRanks, r_max=r_max, j_max=j_max)
plot(x = 0:j_max / j_max, y = densities, type="l")
print(tail(helperTrapezoidRule(densities)))
plot(x = 0:j_max / j_max, y = helperTrapezoidRule(densities), type="l")

### Example 2 ###
j_max <- 12
r_max <- 19
sortedRanks <- c(0,0,1,1,3,5,6,18,19)
densities <- helperGet_X_prima_AB_bounds_bootstrap_for_every_j(sortedRanks=sortedRanks, r_max=r_max, j_max=j_max)
plot(x = 0:j_max / j_max, y = densities, type="l")
print(tail(helperTrapezoidRule(densities)))
plot(x = 0:j_max / j_max, y = helperTrapezoidRule(densities), type="l")
}
\keyword{internal}
