% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{isXlimsValid}
\alias{isXlimsValid}
\title{Check if xlims is a tuple that represents a valid bounded interval in the real space.}
\usage{
isXlimsValid(xlims)
}
\arguments{
\item{xlims}{the tuple to be checked.}
}
\value{
TRUE if it is a valid tuple. Otherwise prints error mesage and returns FALSE
}
\description{
Check if xlims is a tuple that represents a valid bounded interval in the real space.
}
