% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{lowestCommonMultiple}
\alias{lowestCommonMultiple}
\title{Get the lowest common divisor}
\usage{
lowestCommonMultiple(integerArray)
}
\arguments{
\item{integerArray}{an array of integers.}
}
\value{
a integer that is the lcd of the numbers in integerArray
}
\description{
Get the lowest common divisor
}
\keyword{internal}
