% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{mixtureOfUniforms}
\alias{mixtureOfUniforms}
\title{Get a mixture of uniform (AKA tophat) distributions of size 'kernelSize' centered in the points 'kernelPositions'.}
\usage{
mixtureOfUniforms(kernelPositions, kernelSize)
}
\arguments{
\item{kernelPositions}{an array of real values that decribe the positions of the kernels.}

\item{kernelSize}{The size of the uniform kernel: a positive float.}
}
\value{
a single parameter callable function that computes the probability density of the mixture in any point.
}
\description{
Get a mixture of uniform (AKA tophat) distributions of size 'kernelSize' centered in the points 'kernelPositions'.
}
\keyword{internal}
