% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{normalDensity}
\alias{normalDensity}
\title{The probability density function of the normal distribution}
\usage{
normalDensity(mu, sigma)
}
\arguments{
\item{mu}{the mean of the normal distribution.}

\item{sigma}{the standard deviation of the normal distribution.}
}
\value{
Returns a callable function with a single parameter that describes the probability of the normal distribution in that point.
}
\description{
Returns the density function of the normal distribution with mean mu and standard deviation sigma.
The returned function is a single parameter function that returns the probability of the normal distribution in that point.
It is just a convinient wrapper of dnorm from the package 'stat' with some parameter checks.
}
\examples{
dist <- normalDensity(0,1)
dist(0)
}
\seealso{
Other probability density distributions: 
\code{\link{uniformDensity}()}
}
\concept{probability density distributions}
