% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{plot_X_prima_AB}
\alias{plot_X_prima_AB}
\title{Plot the estimated bounds}
\usage{
plot_X_prima_AB(estimated_X_prima_AB_bounds)
}
\arguments{
\item{estimated_X_prima_AB_bounds}{the bounds estimated with \code{\link{get_X_prima_AB_bounds_bootstrap}} or \code{\link{get_X_prima_AB_bounds_DKW}}.}
}
\value{
the ggplot figure object.
}
\description{
retunrs a ggplot2 with the estimated bounds
}
\examples{
### Example 1 ###

X_A_observed <- rnorm(800,mean = 1, sd = 1)
X_B_observed <- rnorm(800,mean = 1.3, sd = 0.5)
res <- get_X_prima_AB_bounds_DKW(X_A_observed, X_B_observed)
densitiesPlot = plot_X_prima_AB(res)
print(densitiesPlot)
}
