% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{xHasEnoughDiffValues}
\alias{xHasEnoughDiffValues}
\title{Check for enough unique values.}
\usage{
xHasEnoughDiffValues(X, EPSILON, minRequiredDiffValues)
}
\arguments{
\item{X}{the array with the values.}

\item{EPSILON}{when will two values be considered different.}

\item{minRequiredDiffValues}{the minimum number of different values required to return TRUE.}
}
\value{
Returns TRUE if the values are OK. FALSE, if there are not enough unique values.
}
\description{
This function checks if there are at least minRequiredDiffValues unique
values in the introduced vector.
}
\examples{
xHasEnoughDiffValues(c(1,2,2,3,1,5,8,9,67,8.5,4,8.3), 1e-9, 6)
}
