% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{xHasEnoughValues}
\alias{xHasEnoughValues}
\title{Check for enough values.}
\usage{
xHasEnoughValues(X, minRequiredValues)
}
\arguments{
\item{X}{the array with the values.}

\item{minRequiredValues}{the minimum number values required to return TRUE.}
}
\value{
Returns TRUE if the values are OK. FALSE, if there are not enough values.
}
\description{
This function checks if there are at least minRequiredValues
values in the introduced vector.
}
\examples{
xHasEnoughValues(c(1,2,2,3,1,5,8,9,67,8.5,4,8.3), 6)
}
