% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVCompare.R
\name{get_X_prima_AB_bounds_bootstrap}
\alias{get_X_prima_AB_bounds_bootstrap}
\title{Estimate X'_A and X'_B bounds with bootstrap 0.632}
\usage{
get_X_prima_AB_bounds_bootstrap(densityX_A, densityX_B, xlims, EPSILON = 0.001)
}
\arguments{
\item{EPSILON}{(optional, default value 1e-20) minimum difference between two values to be considered different.}

\item{X_A_observed}{array of the observed samples (real values) of X_A.}

\item{X_B_observed}{array of the observed samples (real values) of X_B.}

\item{alpha}{(optional, default value 0.2) the error of the confidence interval.}
}
\value{
Returns a list with the following fields:

- p: the values in the interval (0,1) in which the uniform distributions that define the mixtures X'_A and X'_B are centered.

"X_prima_A_density_estimation": the density estimation of X_prima_A,

"X_prima_A_density_upper": the upper bound of the confidence interval of 1 - alpha of the density estimation of X_prima_A

"X_prima_A_density_lower": the lower bound of the confidence interval of 1 - alpha of the density estimation of X_prima_A

"X_prima_B_density_estimation": the density estimation of X_prima_B,

"X_prima_B_density_upper": the upper bound of the confidence interval of 1 - alpha of the density estimation of X_prima_B

"X_prima_B_density_lower": the lower bound of the confidence interval of 1 - alpha of the density estimation of X_prima_B
}
\description{
Estimate the confidence intervals for the density estimations of X'_A and X'_B using bootstrap.
As a bonus, in addition to the density, the bounds of the cumulative density are also compared.
}
\examples{


}
